Don't fall back to unprotected mode if protected is not supported.
Checking the capability on /dev/kvm doesn't work, and in any case if a
client requests a protected VM but it isn't supported we should fail
with an error, not fall back to running it in unprotected mode.

Microdroid clients now need to explicitly choose whether to run in
protected mode or not, the same as raw VM clients.

Test: atest VirtualizationTestCases
Change-Id: If05b871bf83062c1619729ec17eb3342ce5ced30

Former-commit-id: 3994f0059f9916920c6efddec0bf90bff33dc04f