Smoke tests for VM instance secrets
Different instances should have different secrets but within an instance
it should get the same secrets. This should from the very start
of the VM and continue to be true after things have been loaded in the
VM. These tests checks that the microdroid payloads continue to follow
the expectations.

There is a slight chance for the test to fail by chance, but it is
cryptographically insignificant by design.

Bug: 214231981
Test: atest MicrodroidTestApp
Change-Id: I488c7370ad901a73a520bc3fa548d4b9b7959d19

Former-commit-id: 2e6ab799461c3b98888a896c23dfa7bf1b8d95e1