Migrate ComposKeyTestCase to unit tests
Refactor SigningKey to allow it to be decoupled from Dice for
testing. Add unit tests with the same coverage as ComposKeyTestCase,
then remove the latter. (This does mean we are no longer testing the
code inside a VM, but that's covered by ComposTestCase and the end to
end tests.)

As a result we no longer need compos_key_cmd, so remove that. This is
a win as it was becoming a maintenance burden.

Bug: 213891964
Test: atest compsvc_device_tests
Change-Id: If863abcb4e89eeb97a4be6a4a958b691aa1446be

Former-commit-id: 611cc94626825840661ab20f84b3903c48331dec