Don't call onDied() explicitly in MicrodroidTests
This is bad use of the API and isn't needed by any of the tests so get rid
of it. VM will call onDied() when it stop the VM anyway, should a test
need to wait for that (pending b/195765441 for cuttlefish, but we skip
all but one test because of that already). When the test sees onDied()
it lets the test continue as there's no more VM to wait for.

Bug: 220700986
Test: atest MicrodroidTests
Change-Id: I1266f210b960100ebba3ced4d33c9bdd72352575

Former-commit-id: 94ad4c8731541d12d84189b1db589700c3fcc0b6