Only read backing file if chunk index is in range
If the chunk index is out of range (which can be a valid read request),
simply return 0. Previously, we had to deal with fd_server when it
returned an unexpected result. Now we don't need to handle it (the
inconsistant state may still be on Android, but we don't care here).

Bug: 221847239
Test: atest AuthFsHostTest
Change-Id: I3450f496de58f2933fb3818b42e9510c56875061

Former-commit-id: 887b25074a669beb8acbb2ca8b62497e58d30f8f