Use hardcoded pagetable initially.
This pagetable is built into the binary, and maps device memory and
enough RAM for the DT and image. It can later be replaced by a proper
pagetable built in Rust.

Using a hardcoded pagetable means we avoid making any memory access
before enabling it, which avoids a bunch of cache issues.

Bug: 223166344
Test: Ran unprotected VM under crosvm.
Change-Id: I567e377ab93881aa357428f750b11be6b3aad034

Former-commit-id: 267f6c1fcc0015ffe404e20031ecbf697acfa37f